Remove useOverallPageMetricsReport hooks.
